Added Settings Wheel
I added the settings wheel button to the top left side of the
activity_compare_results.xml file. When the wheel is clicked, the user
is sent to the prefs.xml where they are able to type in their search
queries. One issue however is that when you hit the back button, it
sends the user all the way back to the main screen. I also added in a
button back on the activity_compare_results.xml to due manual refreshing
until we can get it to automatically refresh.